Update to 1.8.8
-- Some methods have been renamed
-- Vanilla code now has generics
-- The Forge and FML event buses have been merged
-- Chicken Bones' mods aren't available for 1.8.8 yet

+++ src/main/java/com/choonster/testmod3/block/fluid/BlockFluidNoFlow.java
@@ -18,7 +18,7 @@ public class BlockFluidNoFlow extends BlockFluidFinite {
 	@Override
 	public void updateTick(World world, BlockPos pos, IBlockState state, Random rand) {
 		boolean changed = false;
-		int quantaRemaining = ((Integer) state.getValue(LEVEL)) + 1;
+		int quantaRemaining = state.getValue(LEVEL) + 1;
 
 		// Flow vertically if possible
 		int prevRemaining = quantaRemaining;
